Fix truss integration tests
